+++ pdata/pcommon/spanid_test.go
@@ -20,14 +20,34 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestSpanID(t *testing.T) {
-	sid := EmptySpanID
-	assert.EqualValues(t, [8]byte{}, sid.Bytes())
+func TestNewSpanID(t *testing.T) {
+	sid := SpanID([8]byte{1, 2, 3, 4, 4, 3, 2, 1})
+	assert.Equal(t, [8]byte{1, 2, 3, 4, 4, 3, 2, 1}, [8]byte(sid))
+	assert.False(t, sid.IsEmpty())
+}
+
+func TestNewSpanIDEmpty(t *testing.T) {
+	sid := NewSpanIDEmpty()
+	assert.Equal(t, [8]byte{}, [8]byte(sid))
 	assert.True(t, sid.IsEmpty())
+}
+
+func TestSpanIDHexString(t *testing.T) {
+	sid := SpanID([8]byte{})
 	assert.Equal(t, "", sid.HexString())
 
-	sid = NewSpanID([8]byte{1, 2, 3, 4, 4, 3, 2, 1})
-	assert.EqualValues(t, [8]byte{1, 2, 3, 4, 4, 3, 2, 1}, sid.Bytes())
-	assert.False(t, sid.IsEmpty())
-	assert.Equal(t, "0102030404030201", sid.HexString())
+	sid = SpanID([8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23})
+	assert.Equal(t, "1223ad1223ad1223", sid.HexString())
+}
+
+func TestSpanIDImmutable(t *testing.T) {
+	initialBytes := [8]byte{0x12, 0x23, 0xAD, 0x12, 0x23, 0xAD, 0x12, 0x23}
+	sid := SpanID(initialBytes)
+	assert.Equal(t, SpanID(initialBytes), sid)
+
+	// Get the bytes and try to mutate.
+	sid[4] = 0x89
+
+	// Does not change the already created SpanID.
+	assert.NotEqual(t, SpanID(initialBytes), sid)
 }