+++ pdata/ptrace/json.go
@@ -81,24 +81,10 @@ func readTraceData(iter *jsoniter.Iterator) otlptrace.TracesData {
 
 func readResourceSpans(iter *jsoniter.Iterator) *otlptrace.ResourceSpans {
 	rs := &otlptrace.ResourceSpans{}
-
 	iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
 		switch f {
 		case "resource":
-			iter.ReadObjectCB(func(iter *jsoniter.Iterator, f string) bool {
-				switch f {
-				case "attributes":
-					iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
-						rs.Resource.Attributes = append(rs.Resource.Attributes, json.ReadAttribute(iter))
-						return true
-					})
-				case "droppedAttributesCount", "dropped_attributes_count":
-					rs.Resource.DroppedAttributesCount = json.ReadUint32(iter)
-				default:
-					iter.Skip()
-				}
-				return true
-			})
+			json.ReadResource(iter, &rs.Resource)
 		case "scopeSpans", "scope_spans":
 			iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
 				rs.ScopeSpans = append(rs.ScopeSpans, readScopeSpans(iter))